MWPW-155686 Catalog Milo: Preview Index Follow Up (#391)
With PR #389 we confirmed that the event resource-previewed works in GitHub actions on main repo. The handler for this event is provided now. For the resource path that we get in that event we will

read the resource details
update preview index with these details or
insert the new row with them.
Again, this change cannot be tested on fork. PR needs to be merged and then tested on main repo.

Resolves: MWPW-155686